refactor : ui components
