+++ src/components/header/snb/index.scss
