+++ src/components/notion/NParagraph.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import '@scss/notion.scss';
 import { Caption, RichText, TextBlock, TextItem } from '@types';
-import Linker from '@components/ui/Linker';
+import Linker from '@components/ui/linker';
 
 interface NParagraphProps {
   paragraph?: TextBlock;
