+++ src/components/ui/divider/index.scss
