+++ src/components/ui/float/index.tsx
@@ -1,20 +1,18 @@
 import * as React from 'react';
-import '@scss/components/FloatBox.scss';
+import './index.scss';
 import MoveTopButton from '@components/float/MoveTopButton';
 import SnowflakeHandler from '@components/float/SnowflakeHandler';
 
-interface Props {
+interface FloatBoxProps {
   useTop?: boolean;
   useSnowflake?: boolean;
 }
 
-const FloatBox = ({ useTop, useSnowflake }: Props) => {
+export default function FloatBox({ useTop, useSnowflake }: FloatBoxProps) {
   return (
     <div className={`float-box`}>
       {useTop && <MoveTopButton />}
       {useSnowflake && <SnowflakeHandler />}
     </div>
   );
-};
-
-export default FloatBox;
+}
