+++ src/components/ui/progress/circle/index.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
-import '@scss/components/ui/CircleProgress.scss';
+import './index.scss';
 import { IconCircleProgress } from '@components/icon';
 
 interface CircleProgressProps extends React.HTMLAttributes<HTMLDivElement> {
   height?: number;
 }
 
-const CircleProgress = ({ height }: CircleProgressProps) => {
+export default function CircleProgress({ height }: CircleProgressProps) {
   return (
     <div
       className="circle-progress-box"
@@ -17,6 +17,4 @@ const CircleProgress = ({ height }: CircleProgressProps) => {
       <IconCircleProgress color={'primary'} viewBox={'0 0 2048 2048'} />
     </div>
   );
-};
-
-export default CircleProgress;
+}
