+++ src/components/ui/progress/index.tsx
@@ -1,0 +1,2 @@
+export { default as CircleProgress } from './circle';
+export { default as ScrollProgress } from './scroll';
