+++ src/components/ui/text/index.tsx
@@ -1,0 +1,1 @@
+export { default as HelpText } from './help';
