+++ src/layout/main/index.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import './index.scss';
 import Header from '@module/Header';
 import Footer from '@module/Footer';
-import ScrollProgress from '@components/ui/ScrollProgress';
+import { ScrollProgress } from '@components/ui/progress';
 
 interface Props {
   children: React.ReactNode;
