+++ src/pages/404.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { HeadFC, PageProps } from 'gatsby';
 import SEO from '@components/header/SEO';
-import Linker from '@components/ui/Linker';
+import Linker from '@components/ui/linker';
 import { GlobalPortal } from '@components/GlobalPortal';
 
 const mainStyles = {
